split sendPayment into files
